Updated insertObstacles & insertCoins
Obstacles now look for the previous ground to guarantee the minimum
distance between obstacles and verify if the previous ground had a gap,
if so the first obstacle must be a little further so the player can
jump over it.

Detected a bug with Coins collision. When the player collides with a
Coin, it gets deleted and what we want is to make the Coin disappear!